requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
drf-spectacular==0.27.1
Pillow==10.2.0
openpyxl==3.1.2
django-filter==24.1
